Modified comments api and integrated with post api

+++ quillpost/comments/api/models.py
@@ -4,7 +4,7 @@ from django.db import models
 
 class Comment(models.Model):
     comment_id= models.AutoField(primary_key=True);
-    post_id= models.IntegerField();
+    post_id= models.BigIntegerField();
     username= models.CharField(max_length=255)
     content= models.TextField(max_length=255)
     timestamp= models.DateTimeField(auto_now_add=True)
